lang-switch-agent-demo/pyproject.toml
[project]
name = "lang-switch-agent-demo"
version = "0.1.0"
description = "LiveKit agent that auto-detects user language and switches STT"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
]

[dependency-groups]
dev = []
